RFC add "--cpio-reflink" option for requesting GNU cpio reflinks
The new GNU cpio "--reflink" parameter sees it use copy_file_range()
when copying between source and destination, allowing for copy-on-write
optimization if supported by the underlying filesystem (e.g. Btrfs or
XFS).

When calling cpio with --reflink we should pipe the file list through
padcpio to ensure alignment. This needn't be done when reflinks are
disabled.

Signed-off-by: David Disseldorp <ddiss@suse.de>